pyproject.toml
[tool.poetry]
name = "project-amanda"
version = "0.1.0"
description = ""
authors = ["moorekwon <raccoonhj33@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7, <3.9"
django = "^3.0.8"
djangorestframework = "^3.11.0"
supervisor = "^4.2.0"
boto3 = "^1.14.24"
pillow = "^7.2.0"
awscli = "^1.18.101"
django-storages = "^1.9.1"
gunicorn = "^20.0.4"
django-allauth = "^0.42.0"
markdown = "^3.2.2"
django-filter = "^2.3.0"
dj-rest-auth = "^1.1.0"

[tool.poetry.dev-dependencies]
django-extensions = "^3.0.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
